+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/review/ReviewRegister.java
@@ -2,6 +2,7 @@ package com.triplem.momoim.core.domain.review;
 
 import com.triplem.momoim.core.domain.member.infrastructure.GatheringMemberRepository;
 import com.triplem.momoim.core.domain.review.infrastructure.ReviewRepository;
+import com.triplem.momoim.core.domain.review.model.Review;
 import com.triplem.momoim.exception.BusinessException;
 import com.triplem.momoim.exception.ExceptionCode;
 import lombok.RequiredArgsConstructor;
